+++ service/src/main/kotlin/io/provenance/explorer/service/pricing/fetchers/FetcherConfiguration.kt
@@ -1,6 +1,5 @@
 package io.provenance.explorer.service.pricing.fetchers
 
-import io.provenance.explorer.grpc.flow.FlowApiGrpcClient
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
@@ -8,7 +7,7 @@ import org.springframework.context.annotation.Configuration
 class FetcherConfiguration {
 
     @Bean
-    fun historicalPriceFetcherFactory(flowApiGrpcClient: FlowApiGrpcClient): HistoricalPriceFetcherFactory {
-        return HistoricalPriceFetcherFactory(flowApiGrpcClient)
+    fun historicalPriceFetcherFactory(): HistoricalPriceFetcherFactory {
+        return HistoricalPriceFetcherFactory()
     }
 }