+++ service/src/main/kotlin/io/provenance/explorer/service/pricing/fetchers/HistoricalPriceFetcherFactory.kt
@@ -1,15 +1,12 @@
 package io.provenance.explorer.service.pricing.fetchers
 
 import io.provenance.explorer.config.ExplorerProperties.Companion.UTILITY_TOKEN
-import io.provenance.explorer.grpc.flow.FlowApiGrpcClient
 
-class HistoricalPriceFetcherFactory(
-    private val flowApiGrpcClient: FlowApiGrpcClient
-) {
+class HistoricalPriceFetcherFactory() {
     fun createNhashPricingFetchers(): List<HistoricalPriceFetcher> {
         return listOf(
             OsmosisPriceFetcher(),
-            FlowApiPriceFetcher(UTILITY_TOKEN, listOf("uusd.trading", "uusdc.figure.se", "uusdt.figure.se"), flowApiGrpcClient)
+            NavEventPriceFetcher(UTILITY_TOKEN, listOf("uusd.trading", "uusdc.figure.se", "uusdt.figure.se"))
         )
     }
 }