+++ service/src/main/kotlin/io/provenance/explorer/service/pricing/fetchers/NavEventPriceFetcher.kt
@@ -1,27 +1,25 @@
 package io.provenance.explorer.service.pricing.fetchers
 
+import io.provenance.explorer.domain.entities.NavEventsRecord
 import io.provenance.explorer.domain.models.HistoricalPrice
-import io.provenance.explorer.grpc.flow.FlowApiGrpcClient
 import io.provenance.explorer.service.pricing.utils.HashCalculationUtils
-import io.provlabs.flow.api.NavEvent
 import org.joda.time.DateTime
 
-class FlowApiPriceFetcher(
+class NavEventPriceFetcher(
     val denom: String,
-    val pricingDenoms: List<String>,
-    private val flowApiGrpcClient: FlowApiGrpcClient
+    val pricingDenoms: List<String>
 ) : HistoricalPriceFetcher {
 
     override fun getSource(): String {
-        return "flow-api"
+        return "navevent-table"
     }
     override fun fetchHistoricalPrice(fromDate: DateTime?): List<HistoricalPrice> {
-        val onChainNavEvents = getMarkerNavByPriceDenoms(fromDate)
+        val onChainNavEvents = NavEventsRecord.getNavEvents(denom = denom, priceDenoms = pricingDenoms, fromDate = fromDate)
         return onChainNavEvents.map { navEvent ->
             val volumeHash = HashCalculationUtils.calculateVolumeHash(navEvent.volume)
-            val pricePerHash = HashCalculationUtils.getPricePerHashFromMicroUsd(navEvent.priceAmount, navEvent.volume)
+            val pricePerHash = HashCalculationUtils.getPricePerHashFromMicroUsd(navEvent.priceAmount!!, navEvent.volume)
             HistoricalPrice(
-                time = navEvent.blockTime,
+                time = navEvent.blockTime.millis / 1000,
                 high = pricePerHash,
                 low = pricePerHash,
                 close = pricePerHash,
@@ -30,9 +28,5 @@ class FlowApiPriceFetcher(
                 source = getSource()
             )
         }
-    }
-
-    fun getMarkerNavByPriceDenoms(fromDate: DateTime?): List<NavEvent> {
-        return flowApiGrpcClient.getAllMarkerNavByPriceDenoms(denom = denom, priceDenoms = pricingDenoms, fromDate = fromDate)
     }
 }